requirements.txt
aiohttp==3.0.9
async-timeout==2.0.1
attrs==17.4.0
chardet==3.0.4
idna==2.6
idna-ssl==1.0.1
multidict==4.1.0
numpy==1.14.2
Pillow==5.0.0
uvloop==0.9.1
yarl==1.1.1
